+++ plugins/telescope.lua
@@ -1,0 +1,8 @@
+return {
+  {
+    "nvim-telescope/telescope-file-browser.nvim",
+    dependencies = { "nvim-telescope/telescope.nvim", "nvim-lua/plenary.nvim" },
+    config = function() require("telescope").load_extension "file_browser" end,
+    event = "User AstroFile",
+  },
+}